extract out more complicated components
