+++ src/renderer/src/routes/app/projects/$projectId/index.tsx
@@ -1,53 +1,23 @@
-import {
-	Suspense,
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	type FocusEvent,
-	type MouseEvent,
-	type ReactNode,
-} from 'react'
-import {
-	useDocumentCreatedBy,
-	useIconUrl,
-	useManyDocs,
-	useOwnDeviceInfo,
-	useOwnRoleInProject,
-	useProjectSettings,
-} from '@comapeo/core-react'
+import { Suspense } from 'react'
+import { useOwnRoleInProject, useProjectSettings } from '@comapeo/core-react'
 import Box from '@mui/material/Box'
 import CircularProgress from '@mui/material/CircularProgress'
 import Divider from '@mui/material/Divider'
-import List from '@mui/material/List'
-import ListItemButton from '@mui/material/ListItemButton'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 import { alpha } from '@mui/material/styles'
-import { useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
-import { useVirtualizer } from '@tanstack/react-virtual'
 import { defineMessages, useIntl } from 'react-intl'
 
-import {
-	BLACK,
-	BLUE_GREY,
-	COMAPEO_BLUE,
-	DARK_GREY,
-	LIGHT_COMAPEO_BLUE,
-	LIGHT_GREY,
-	WHITE,
-} from '../../../../colors'
+import { BLACK, DARK_GREY, LIGHT_GREY } from '../../../../colors'
 import { Icon } from '../../../../components/icon'
-import { ButtonLink, TextLink } from '../../../../components/link'
+import { ButtonLink } from '../../../../components/link'
 import {
 	COMAPEO_CORE_REACT_ROOT_QUERY_KEY,
 	COORDINATOR_ROLE_ID,
 	CREATOR_ROLE_ID,
-	getMatchingCategoryForObservation,
-	getMatchingCategoryForTrack,
 } from '../../../../lib/comapeo'
-import { getLocaleStateQueryOptions } from '../../../../lib/queries/app-settings'
+import { DisplayedDataList } from './-displayed-data/list'
 
 export const Route = createFileRoute('/app/projects/$projectId/')({
 	loader: async ({ context, params }) => {
@@ -191,434 +161,10 @@ function RouteComponent() {
 						</Box>
 					}
 				>
-					<ListedDataSection projectId={projectId} />
+					<DisplayedDataList projectId={projectId} />
 				</Suspense>
 			</Box>
 		</Stack>
-	)
-}
-
-function ListedDataSection({ projectId }: { projectId: string }) {
-	const { formatMessage: t, formatDate } = useIntl()
-
-	const { highlightedDocument } = Route.useSearch()
-	const navigate = Route.useNavigate()
-
-	const { data: lang } = useSuspenseQuery({
-		...getLocaleStateQueryOptions(),
-		select: ({ value }) => value,
-	})
-
-	const { data: observations } = useManyDocs({
-		projectId,
-		docType: 'observation',
-		lang,
-	})
-
-	const { data: tracks } = useManyDocs({
-		projectId,
-		docType: 'track',
-		lang,
-	})
-
-	const { data: categories } = useManyDocs({
-		projectId,
-		docType: 'preset',
-		lang,
-	})
-
-	const observationsWithCategory = useMemo(() => {
-		return observations.map((o) => ({
-			type: 'observation' as const,
-			value: o,
-			category: getMatchingCategoryForObservation(o.tags, categories),
-		}))
-	}, [observations, categories])
-
-	const tracksWithCategory = useMemo(() => {
-		return tracks.map((t) => ({
-			type: 'track' as const,
-			value: t,
-			category: getMatchingCategoryForTrack(t, categories),
-		}))
-	}, [tracks, categories])
-
-	const sortedListData = useMemo(() => {
-		return [...observationsWithCategory, ...tracksWithCategory].sort((a, b) => {
-			return a.value.createdAt < b.value.createdAt ? 1 : -1
-		})
-	}, [observationsWithCategory, tracksWithCategory])
-
-	const listRef = useRef<HTMLUListElement | null>(null)
-
-	const onFocus = useCallback(
-		(event: FocusEvent<HTMLUListElement>) => {
-			const el = event.target.closest('[data-docid]')
-
-			if (!(el instanceof HTMLElement)) {
-				return
-			}
-
-			const dataType = el.getAttribute('data-datatype')
-			const docId = el.getAttribute('data-docid')
-
-			if (!(dataType && docId)) {
-				return
-			}
-
-			if (!(dataType === 'observation' || dataType === 'track')) {
-				return
-			}
-
-			if (highlightedDocument && docId === highlightedDocument.docId) {
-				return
-			}
-
-			navigate({
-				search: {
-					highlightedDocument: {
-						type: dataType,
-						docId,
-					},
-				},
-			})
-		},
-		[navigate, highlightedDocument],
-	)
-
-	const onMouseMove = useCallback((event: MouseEvent<HTMLUListElement>) => {
-		const el = (event.target as HTMLElement).closest('[data-docid]')
-
-		if (!(el instanceof HTMLElement)) {
-			return
-		}
-
-		// NOTE: We defer to the onFocus callback in order to determine the navigation changes.
-		el.focus()
-	}, [])
-
-	const getItemKey = useCallback(
-		(index: number) => {
-			const item = sortedListData[index]
-
-			// Shouldn't happen but fail loudly if it does
-			if (!item) {
-				throw new Error(`Could not get item from listed data at index ${index}`)
-			}
-
-			return item.value.docId
-		},
-		[sortedListData],
-	)
-
-	const rowVirtualizer = useVirtualizer({
-		count: sortedListData.length,
-		getScrollElement: () => listRef.current,
-		estimateSize: () => 100,
-		getItemKey,
-		overscan: 10,
-	})
-
-	const isMouseOverListRef = useRef(false)
-
-	const itemIndexToScrollTo = highlightedDocument?.docId
-		? sortedListData.findIndex(
-				({ value }) => value.docId === highlightedDocument.docId,
-			)
-		: undefined
-
-	const { scrollToIndex } = rowVirtualizer
-
-	useEffect(() => {
-		// Prevents issues with manually scrolling on the list due to the behavior
-		// implemented in onMouseMove
-		if (isMouseOverListRef.current) {
-			return
-		}
-
-		if (itemIndexToScrollTo) {
-			scrollToIndex(itemIndexToScrollTo, {
-				align: 'center',
-			})
-		}
-	}, [itemIndexToScrollTo, scrollToIndex])
-
-	return sortedListData.length > 0 ? (
-		<List
-			component="ul"
-			ref={listRef}
-			disablePadding
-			onFocus={onFocus}
-			onMouseMove={onMouseMove}
-			onMouseEnter={() => {
-				isMouseOverListRef.current = true
-			}}
-			onMouseLeave={() => {
-				isMouseOverListRef.current = false
-			}}
-			sx={{ overflow: 'auto', scrollbarColor: 'initial' }}
-		>
-			<Box
-				position="relative"
-				height={`${rowVirtualizer.getTotalSize()}px`}
-				width="100%"
-			>
-				{rowVirtualizer.getVirtualItems().map((row) => {
-					const { type, category, value } = sortedListData[row.index]!
-
-					return (
-						<Box
-							key={row.key}
-							sx={{
-								position: 'absolute',
-								top: 0,
-								left: 0,
-								width: '100%',
-								height: `${row.size}px`,
-								transform: `translateY(${row.start}px)`,
-							}}
-						>
-							<ListItemButton
-								key={value.docId}
-								data-datatype={type}
-								data-docid={value.docId}
-								disableGutters
-								disableTouchRipple
-								selected={highlightedDocument?.docId === value.docId}
-								onClick={() => {
-									if (type === 'observation') {
-										navigate({
-											to: './observations/$observationDocId',
-											params: { observationDocId: value.docId },
-										})
-									} else {
-										navigate({
-											to: './tracks/$trackDocId',
-											params: { trackDocId: value.docId },
-										})
-									}
-								}}
-								sx={{ borderBottom: `1px solid ${LIGHT_GREY}`, padding: 4 }}
-							>
-								<Suspense>
-									<SyncedIndicatorLine
-										projectId={projectId}
-										originalVersionId={value.originalVersionId}
-									/>
-								</Suspense>
-								<Stack
-									direction="row"
-									flex={1}
-									useFlexGap
-									gap={2}
-									overflow="auto"
-								>
-									<Stack
-										direction="column"
-										flex={1}
-										justifyContent="center"
-										overflow="hidden"
-									>
-										<Typography
-											fontWeight={500}
-											textOverflow="ellipsis"
-											whiteSpace="nowrap"
-											overflow="hidden"
-										>
-											{category?.name ||
-												t(
-													type === 'observation'
-														? m.observationCategoryNameFallback
-														: m.trackCategoryNameFallback,
-												)}
-										</Typography>
-
-										<Typography
-											textOverflow="ellipsis"
-											whiteSpace="nowrap"
-											overflow="hidden"
-										>
-											{formatDate(value.createdAt, {
-												year: 'numeric',
-												month: 'short',
-												day: '2-digit',
-												minute: '2-digit',
-												hour: '2-digit',
-												hourCycle: 'h12',
-											})}
-										</Typography>
-									</Stack>
-
-									<Box
-										display="flex"
-										justifyContent="center"
-										alignItems="center"
-									>
-										{category?.iconRef?.docId ? (
-											<Suspense
-												fallback={
-													<Box
-														display="flex"
-														justifyContent="center"
-														alignItems="center"
-														height={48}
-														width={48}
-													>
-														<CircularProgress disableShrink size={30} />
-													</Box>
-												}
-											>
-												<DisplayedCategoryAndAttachments
-													projectId={projectId}
-													categoryName={category.name}
-													borderColor={category.color || BLUE_GREY}
-													iconDocumentId={category.iconRef.docId}
-												/>
-											</Suspense>
-										) : (
-											<CategoryIconContainer borderColor={BLUE_GREY}>
-												<Icon name="material-place" size={40} />
-											</CategoryIconContainer>
-										)}
-									</Box>
-								</Stack>
-							</ListItemButton>
-						</Box>
-					)
-				})}
-			</Box>
-		</List>
-	) : (
-		<AddObservationsCard projectId={projectId} />
-	)
-}
-
-function SyncedIndicatorLine({
-	projectId,
-	originalVersionId,
-}: {
-	projectId: string
-	originalVersionId: string
-}) {
-	const { data: ownDeviceInfo } = useOwnDeviceInfo()
-
-	const { data: createdBy } = useDocumentCreatedBy({
-		projectId,
-		originalVersionId,
-	})
-
-	return ownDeviceInfo.deviceId !== createdBy ? (
-		<Box
-			position="absolute"
-			width={8}
-			left={0}
-			bottom={0}
-			top={0}
-			bgcolor={COMAPEO_BLUE}
-		/>
-	) : null
-}
-
-// TODO: Display attachments
-function DisplayedCategoryAndAttachments({
-	borderColor,
-	categoryName,
-	projectId,
-	iconDocumentId,
-}: {
-	borderColor: string
-	categoryName: string
-	projectId: string
-	iconDocumentId: string
-}) {
-	const { formatMessage: t } = useIntl()
-
-	const { data: iconURL } = useIconUrl({
-		projectId,
-		iconId: iconDocumentId,
-		mimeType: 'image/png',
-		size: 'small',
-		pixelDensity: 3,
-	})
-
-	return (
-		<CategoryIconContainer borderColor={borderColor}>
-			<img
-				src={iconURL}
-				alt={t(m.categoryIconAlt, { name: categoryName })}
-				style={{ aspectRatio: 1, maxHeight: 48 }}
-			/>
-		</CategoryIconContainer>
-	)
-}
-
-function CategoryIconContainer({
-	borderColor,
-	children,
-}: {
-	borderColor: string
-	children: ReactNode
-}) {
-	return (
-		<Box
-			display="flex"
-			alignItems="center"
-			justifyContent="center"
-			padding={2}
-			overflow="hidden"
-			borderRadius="50%"
-			border={`3px solid ${borderColor}`}
-		>
-			{children}
-		</Box>
-	)
-}
-
-function AddObservationsCard({ projectId }: { projectId: string }) {
-	const { formatMessage: t } = useIntl()
-
-	return (
-		<Box display="flex" flex={1} padding={6}>
-			<Stack
-				direction="column"
-				useFlexGap
-				gap={4}
-				alignItems="center"
-				borderRadius={2}
-				border={`1px solid ${BLUE_GREY}`}
-				paddingX={6}
-				paddingY={10}
-				justifyContent="center"
-				flex={1}
-			>
-				<Box
-					borderRadius="100%"
-					bgcolor={LIGHT_COMAPEO_BLUE}
-					display="flex"
-					justifyContent="center"
-					alignItems="center"
-					padding={6}
-				>
-					<Icon name="comapeo-cards" htmlColor={WHITE} size={40} />
-				</Box>
-
-				<Typography variant="h1" textAlign="center" fontWeight={500}>
-					{t(m.addObservationsTitle)}
-				</Typography>
-
-				<Typography textAlign="center" fontWeight={400}>
-					{t(m.addObservationsDescription)}
-				</Typography>
-
-				<TextLink
-					underline="none"
-					to="/app/projects/$projectId/exchange"
-					params={{ projectId }}
-				>
-					{t(m.goToExchange)}
-				</TextLink>
-			</Stack>
-		</Box>
 	)
 }
 
@@ -647,38 +193,5 @@ const m = defineMessages({
 		id: 'routes.app.projects.$projectId.index.invite',
 		defaultMessage: 'Invite',
 		description: 'Link text to navigate to invite collaborators page.',
-	},
-	addObservationsTitle: {
-		id: 'routes.app.projects.$projectId.index.addObservationsTitle',
-		defaultMessage: 'Add Observations',
-		description:
-			'Title of card that is displayed when project has no observations to display.',
-	},
-	addObservationsDescription: {
-		id: 'routes.app.projects.$projectId.index.addObservationsDescription',
-		defaultMessage: 'Use Exchange to add Collaborator Observations',
-		description:
-			'Description of card that is displayed when project has no observations to display.',
-	},
-	goToExchange: {
-		id: 'routes.app.projects.$projectId.index.goToExchange',
-		defaultMessage: 'Go to Exchange',
-		description: 'Link text to navigate to Exchange page.',
-	},
-	observationCategoryNameFallback: {
-		id: 'routes.app.projects.$projectId.index.observationCategoryNameFallback',
-		defaultMessage: 'Observation',
-		description: 'Fallback name for observation without a matching category.',
-	},
-	trackCategoryNameFallback: {
-		id: 'routes.app.projects.$projectId.index.trackCategoryNameFallback',
-		defaultMessage: 'Track',
-		description: 'Fallback name for track without a matching category.',
-	},
-	categoryIconAlt: {
-		id: 'routes.app.projects.$projectId.index.categoryIconAlt',
-		defaultMessage: 'Icon for {name} category',
-		description:
-			'Alt text for icon image displayed for category (used for accessibility tools).',
 	},
 })
